requirements.txt
google-cloud-aiplatform
google-cloud-storage
google-genai
gitpython
google-adk
streamlit
python-dotenv